move image test case to utils
